Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,8 +2,8 @@
 
 int main() 
 {
-    float a,b,c;
-    scanf("%f %f %f",&a,&b,&c);
-    printf("Average: %.2f", (a+b+c/3));
+    int a,b,c;
+    scanf("%d %d %d",&a,&b,&c);
+    printf("Average: %.2f", a+b+c/3);
     return 0;
 }
